component restructure, add page for application list and confirmation page
