+++ my-app/src/components/PersonalDetails.js
@@ -6,8 +6,15 @@ const PersonalDetails = () => {
       <div className="border-b border-gray-900/10 pb-12">
         <h2 className="text-base/7 font-semibold text-gray-900">Personal Information</h2>
         <p className="mt-1 text-sm/6 text-gray-600">Use a permanent address where you can receive mail.</p>
+        
 
         <div className="mt-10 grid grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6">
+        <div className="sm:col-span-2">
+            <label htmlFor="last-name" className="block text-sm/6 font-medium text-gray-900">Last name / Nom</label>
+            <div className="mt-2">
+              <input type="text" name="last-name" id="last-name" autoComplete="family-name" className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm/6" />
+            </div>
+          </div>
           <div className="sm:col-span-2">
             <label htmlFor="first-name" className="block text-sm/6 font-medium text-gray-900">First name</label>
             <div className="mt-2">
@@ -23,11 +30,20 @@ const PersonalDetails = () => {
           </div>
 
           <div className="sm:col-span-2">
-            <label htmlFor="last-name" className="block text-sm/6 font-medium text-gray-900">Last name</label>
+            <label htmlFor="first-name" className="block text-sm/6 font-medium text-gray-900">First name</label>
             <div className="mt-2">
-              <input type="text" name="last-name" id="last-name" autoComplete="family-name" className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm/6" />
+              <input type="text" name="first-name" id="first-name" autoComplete="given-name" className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm/6" />
             </div>
           </div>
+
+          <div className="sm:col-span-2">
+            <label htmlFor="middle-name" className="block text-sm/6 font-medium text-gray-900">Middle name</label>
+            <div className="mt-2">
+              <input type="text" name="middle-name" id="middle-name" autoComplete="middle-name" className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm/6" />
+            </div>
+          </div>
+
+      
 
           <div className="sm:col-span-4">
             <label htmlFor="email" className="block text-sm/6 font-medium text-gray-900">Email address</label>
